backend/requirements.txt
# Butler Backend - Production Dependencies
fastapi>=0.117.0
uvicorn[standard]>=0.36.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
email-validator>=2.0.0
python-multipart>=0.0.9
starlette>=0.27.0
pyjwt[crypto]>=2.8.0
argon2-cffi>=23.1.0
cryptography>=42.0.0
webauthn>=2.2.0
authlib>=1.3.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.2.5
psycopg2-binary>=2.9.9
redis>=5.0.0
apscheduler>=3.10.4
structlog>=24.1.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
opentelemetry-instrumentation>=0.45b0
trafilatura>=1.9.0
beautifulsoup4>=4.12.0
jsonschema>=4.21.0
typing-extensions>=4.9.0
numpy>=1.26.0
websockets>=16.0
httpx>=0.27.0
neo4j>=5.14.0
qdrant-client>=1.7.0
openai>=1.12.0
psutil>=7.2.2
boto3>=1.34.0
botocore>=1.34.0
aiokafka>=0.9.0
prometheus-client>=0.25.0
tenacity>=8.2.0
noisereduce>=3.0.0
soundfile>=0.12.0
onnxruntime>=1.17.0
ffmpeg-python>=0.2.0
pyacoustid>=1.3.0
youtube-transcript-api>=0.6.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# LangChain/LangGraph for orchestration
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
langchain-mcp-adapters>=0.1.0

# Hermes agent dependencies
fire>=0.7.1,<1
python-dotenv>=1.2.1,<2
rich>=14.3.3,<15
tenacity>=9.1.4,<10
pyyaml>=6.0.2,<7
jinja2>=3.1.5,<4
prompt_toolkit>=3.0.52,<4